remove bin file for CLI as per convo, and revert work from yesterday
